+++ appPlugins/ToolNCC.py
@@ -1057,6 +1057,7 @@ class NonCopperClear(AppTool, Gerber):
 
         min_list = list(min_dict.keys())
         min_dist = min(min_list)
+        min_dist -= 0.0001  # make sure that this works for isolation case
 
         return msg, min_dist